Make dotplot assembly loading fail
